Bugs Corrigidos (12/02/2023)

+++ functions.py
@@ -12,8 +12,13 @@ class Functions:
         self.delta = self.b ** 2 - (4 * self.a * self.c)
         
         if self.delta < 0:
+            try:
+                os.remove("Passo a passo.txt")
+                body_txt(self.a, self.b, self.c, self.delta, 0, 0, 0, 0)
+            except:
+                body_txt(self.a, self.b, self.c, self.delta, 0, 0, 0, 0)
             return f"Delta Negativo (D = {self.delta})"
-        
+            
         self.division = 2 * self.a
         
         self.raiz_delta = pow(self.delta, 1/2)
@@ -23,16 +28,19 @@ class Functions:
         
         self.x2 = -self.b - self.raiz_delta
         self.x2 /= self.division
-
+        
         return f"X1 = {self.x1}, X2 = {self.x2}"
 
     def step(self, a, b,c):
-        try:
-            os.remove("Passo a passo.txt")
-            body_txt(a,b,c, self.delta, self.raiz_delta, self.division, self.x1, self.x2)
-        except:
-            body_txt(a,b, c, self.delta, self.raiz_delta, self.division, self.x1, self.x2)
-
+        
+        if self.delta >=0:
+            try:
+                os.remove("Passo a passo.txt")
+                body_txt(a,b,c, self.delta, self.raiz_delta, self.division, self.x1, self.x2)
+            except:
+                body_txt(a,b, c, self.delta, self.raiz_delta, self.division, self.x1, self.x2)
+        else:
+            pass
